[front] feat: Add bounding boxes to pptx inspect
